+++ indico/modules/events/contributions/util.py
@@ -36,6 +36,7 @@ from indico.util.spreadsheets import csv_text_io_wrapper
 from indico.util.string import format_email_with_name, validate_email
 from indico.web.flask.templating import get_template_module
 from indico.web.flask.util import send_file, url_for
+from indico.web.forms.base import IndicoForm
 from indico.web.util import jsonify_data
 
 
@@ -177,7 +178,7 @@ def generate_spreadsheet_from_contributions(contributions):
     return headers, rows
 
 
-def make_contribution_form(event, *, management=True):
+def make_contribution_form(event, *, management=True, only_custom_fields=False):
     """Extend the contribution WTForm to add the extra fields.
 
     Each extra field will use a field named ``custom_ID``.
@@ -187,7 +188,8 @@ def make_contribution_form(event, *, management=True):
     """
     from indico.modules.events.contributions.forms import ContributionForm
 
-    form_class = type('_ContributionForm', (ContributionForm,), {})
+    base_class = IndicoForm if only_custom_fields else ContributionForm
+    form_class = type('_ContributionForm', (base_class,), {})
     for custom_field in event.contribution_fields:
         field_impl = custom_field.mgmt_field if management else custom_field.field
         if field_impl is None: